refactor(service): update service interfaces ICustomerService, IOrderService, IProductService
