+++ src/main/java/com/myroslav/cosmickitties/service/abstraction/IOrderService.java
@@ -1,9 +1,10 @@
 package com.myroslav.cosmickitties.service.abstraction;
 
 import com.myroslav.cosmickitties.dto.OrderDTO;
+
 import java.util.List;
 
-public interface OrderService {
+public interface IOrderService {
     OrderDTO create(OrderDTO dto);
     OrderDTO getById(Long id);
     List<OrderDTO> getAll();
