+++ src/main/java/com/myroslav/cosmickitties/service/abstraction/IProductService.java
@@ -1,4 +1,4 @@
-package com.myroslav.cosmickitties.service;
+package com.myroslav.cosmickitties.service.abstraction;
 
 import com.myroslav.cosmickitties.dto.ProductDTO;
